Color added + CSS added
There is a slight change within the HTML to make it look the other code that has been written. Additionally, I also added the CSS with the same lay-out to make it more repetitive throughout the file. It's important to use the same lay-out so its easier to change and look into for others.

What to do: Fix the button-container.